pyproject.toml
[project]
name = "linkedin-env"
version = "0.1.0"
description = "LinkedIn content strategy RL environment for the Meta PyTorch OpenEnv Hackathon"
requires-python = ">=3.11"
dependencies = [
    "openenv-core>=0.2.3",
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.29.0",
    "openai>=1.30.0",
    "pydantic>=2.6.0",
    "numpy>=1.26.0",
    "httpx>=0.27.0",
    "websockets>=12.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project.scripts]
server = "server.app:main"

[tool.hatch.build.targets.wheel]
packages = ["env"]
